backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
python-dotenv==1.0.1
gunicorn==22.0.0
requests==2.32.3
feedparser==6.0.11
beautifulsoup4==4.12.3
tqdm==4.66.4
google-search-results==2.4.2
